Hack for Super Micro serialnumber
history:  Super Micro boards return a fact for 'serialnumber' of 1234567890' for all boards.  Java converts this into an INT (logically), but it's cast as a String, so rundeck chokes.  This hack allows Super Micro boards to work, but hides the serial numbers of all devices to Rundeck.  This does NOT change the underlying fact.